pyproject.toml
[project]
name = "osticket-api"
version = "0.3.0"
description = "A new API for OSTicket"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "pymysql>=1.1.2",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "python-dotenv>=1.0.0",
    "pytest>=9.0.2",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0"
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
addopts = """
    --cov=main --cov-report=term-missing --cov-report=html
"""
